requirements/project-requirements.txt
celery==5.6.2

# if you change this, update the documentation
elasticsearch==9.2.1

Django==5.2.10
djangorestframework==3.16.1
django-rest-email-auth==5.0.0
django-ses==4.6.0

psycopg2-binary==2.9.11

certego-saas==0.7.11
slack-sdk==3.39.0

uwsgitop==0.12
uwsgi==2.0.31

joblib==1.5.3
pandas==3.0.0
scikit-learn==1.8.0
numpy==2.4.1
datasketch==1.9.0
